refactor : 자동차 경주 시뮬레이션 CarSimulator 클래스의 패키지 변경
